+++ ClientApp/src/services/index.js
@@ -9,3 +9,4 @@ export * from './role-permissions.service';
 export * from './permission-groups.service';
 export * from './permission-group-mappings.service';
 export * from './role-permission-group-mappings.service';
+export * from './user-access-levels.service';