+++ Levendr/Plugins/BasePlugin.cs
@@ -1,0 +1,7 @@
+namespace Levendr.Plugins
+{
+    public class BasePlugin
+    {
+        
+    }
+}